move stomp stuff into snapshotSaveLoad namespace
and out of exceptions.h, where they do not belong
