+++ include/clasp/core/exceptions.h
@@ -209,20 +209,6 @@ public:
   int getExitResult() { return this->_ExitResult; };
 };
 
-typedef enum { undef, stomp, noStomp, testStomp } ForwardingEnum;
-
-
-struct SaveLispAndDie {
-  string _FileName;
-  bool _Executable;
-  string _LibDir;
-  bool   _Exit;  // Set to true unless debugging
-  ForwardingEnum _ForwardingKind;
-  bool           _TestMemory;
-  SaveLispAndDie(const std::string& filename, bool executable, const std::string& libDir, bool ep=true, ForwardingEnum fk=noStomp, bool tm=true)
-      : _FileName(filename), _Executable(executable), _LibDir(libDir), _Exit(ep), _ForwardingKind(fk), _TestMemory(tm) {};
-};
-
 /*! To exit the program throw this exception
  */
 class TerminateProgramIfBatch {
